backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-generativeai==0.8.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
sse-starlette==2.1.3
